# Conflicts:
#	activation/activation.go
#	activation/activation_test.go
#	activation/activationdb_test.go
#	activation/atxdb.go
#	api/grpcserver/grpcserver_test.go
#	api/grpcserver/mesh_service.go
#	blocks/blockoracle_test.go
#	cmd/node/app_test.go
#	cmd/node/multi_node.go
#	cmd/node/node.go
#	hare/eligibility/oracle.go
#	hare/eligibility/oracle_test.go
#	mesh/reward_test.go

+++ blocks/blockeligibilityvalidator.go
@@ -17,7 +17,7 @@ type blockDB interface {
 // BlockEligibilityValidator holds all the dependencies for validating block eligibility.
 type BlockEligibilityValidator struct {
 	committeeSize  uint32
-	layersPerEpoch uint16
+	layersPerEpoch uint32
 	activationDb   activationDB
 	blocks         blockDB
 	beaconProvider BeaconGetter
@@ -27,7 +27,7 @@ type BlockEligibilityValidator struct {
 
 // NewBlockEligibilityValidator returns a new BlockEligibilityValidator.
 func NewBlockEligibilityValidator(
-	committeeSize uint32, layersPerEpoch uint16, activationDb activationDB, beaconProvider BeaconGetter,
+	committeeSize uint32, layersPerEpoch uint32, activationDb activationDB, beaconProvider BeaconGetter,
 	validateVRF VRFValidationFunction, blockDB blockDB, log log.Log) *BlockEligibilityValidator {
 
 	return &BlockEligibilityValidator{
